refactor(recall)!: remove auto-recall PostToolUse hook in favor of teamai-recall subagent
auto-recall silently fired on every Bash/Grep/WebSearch/WebFetch tool call to
search team knowledge, but was noisy and low-hit-rate in practice, and its job
is now done better by the teamai-recall subagent (proactive, codebase-graph
aware, structured output). Installed machines self-clean the old hook entries
on the next `teamai pull` / `hooks inject` via the existing reconcile engine.

contribute-check's knowledge-gap detection (Phase 2) depended on auto-recall's
quality cache, so `teamai recall` now records recall quality itself (moved
into a standalone recall-quality.ts module) — the feature keeps working, just
with a new, more accurate data source (manual recall + the recall subagent's
underlying command, instead of a passive hook).
