+++ src/__tests__/hook-handlers.test.ts
@@ -9,7 +9,6 @@ const mockParseHookEvent = vi.fn().mockResolvedValue({ type: 'session_start', ti
 const mockAppendEvent = vi.fn().mockResolvedValue(undefined);
 const mockTrackFromParsed = vi.fn().mockResolvedValue(undefined);
 const mockTrackSlashFromParsed = vi.fn().mockResolvedValue(undefined);
-const mockAutoRecallFromParsed = vi.fn().mockResolvedValue(null);
 const mockContributeCheckForSession = vi.fn().mockResolvedValue({ hint: null });
 const mockDoUpdate = vi.fn().mockResolvedValue(undefined);
 
@@ -32,14 +31,6 @@ vi.mock('../usage-tracker.js', () => ({
   appendUsageEvent: vi.fn().mockResolvedValue(undefined),
   updateKnownSkills: vi.fn().mockResolvedValue(undefined),
 }));
-
-vi.mock('../auto-recall.js', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('../auto-recall.js')>();
-  return {
-    ...actual,
-    autoRecallFromInput: mockAutoRecallFromParsed,
-  };
-});
 
 vi.mock('../contribute-check.js', () => ({
   contributeCheck: vi.fn().mockResolvedValue(undefined),
@@ -115,40 +106,12 @@ describe('hook-handlers registry', () => {
     expect(skillHandlers).toContain('track');
   });
 
-  it('post-tool-use Bash/Grep/WebSearch/WebFetch have auto-recall', () => {
+  it('post-tool-use Bash/Grep/WebSearch/WebFetch have no registered handlers', () => {
     const registry = buildHandlerRegistry();
     for (const matcher of ['Bash', 'Grep', 'WebSearch', 'WebFetch']) {
-      const handlers = registry
-        .filter((r) => r.event === 'post-tool-use' && r.matcher === matcher)
-        .map((r) => r.handler.name);
-      expect(handlers).toContain('auto-recall');
+      const handlers = registry.filter((r) => r.event === 'post-tool-use' && r.matcher === matcher);
+      expect(handlers).toHaveLength(0);
     }
-  });
-
-  it('auto-recall handler parses stdin and delegates to autoRecallFromInput', async () => {
-    const registry = buildHandlerRegistry();
-    const bashHandler = registry.find(
-      (r) => r.event === 'post-tool-use' && r.matcher === 'Bash',
-    )!.handler;
-
-    const stdin = {
-      tool_name: 'Bash',
-      tool_input: { command: 'npm test' },
-      tool_output: 'Error: module not found',
-      session_id: 'session-123',
-    };
-    const expectedOutput = JSON.stringify({ hookSpecificOutput: { additionalContext: 'context' } });
-    mockAutoRecallFromParsed.mockResolvedValueOnce(expectedOutput);
-
-    const result = await bashHandler.execute(stdin, 'claude');
-
-    expect(mockAutoRecallFromParsed).toHaveBeenCalledWith({
-      toolName: 'Bash',
-      toolInput: { command: 'npm test' },
-      toolOutput: 'Error: module not found',
-      sessionId: 'session-123',
-    });
-    expect(result).toBe(expectedOutput);
   });
 
   it('prompt-submit has track-slash and dashboard-report', () => {
