+++ src/__tests__/hooks-dispatch-format.test.ts
@@ -86,8 +86,8 @@ describe('hooks — merged dispatch format', () => {
         totalEntries += matchers.length;
       }
 
-      // Should have: SessionStart(1) + Stop(1) + PostToolUse(Skill:1, *:1, TodoWrite:1, Bash:1, Grep:1, WebSearch:1, WebFetch:1) + UserPromptSubmit(1) = 10
-      expect(totalEntries).toBeLessThanOrEqual(10);
+      // Should have: SessionStart(1) + Stop(1) + PostToolUse(*:1, Skill:1, TodoWrite:1) + UserPromptSubmit(1) = 6
+      expect(totalEntries).toBeLessThanOrEqual(6);
     });
 
     it('SessionStart has exactly one entry with wildcard matcher', async () => {
@@ -116,7 +116,7 @@ describe('hooks — merged dispatch format', () => {
       const result = mockFiles['/test/settings.json'] as { hooks: Record<string, ClaudeHookMatcher[]> };
       const postToolUse = result.hooks.PostToolUse;
       const matchers = postToolUse.map((m: ClaudeHookMatcher) => m.matcher).sort();
-      expect(matchers).toEqual(['*', 'Bash', 'Grep', 'Skill', 'TodoWrite', 'WebFetch', 'WebSearch']);
+      expect(matchers).toEqual(['*', 'Skill', 'TodoWrite']);
     });
 
     it('UserPromptSubmit has exactly one entry with wildcard matcher', async () => {
@@ -141,8 +141,8 @@ describe('hooks — merged dispatch format', () => {
       await injectHooks('/test/settings.json', 'claude');
 
       const result = mockFiles['/test/settings.json'] as { hooks: Record<string, ClaudeHookMatcher[]> };
-      const bashEntry = result.hooks.PostToolUse.find((m: ClaudeHookMatcher) => m.matcher === 'Bash');
-      expect(bashEntry!.hooks[0].command).toContain('--matcher Bash');
+      const skillEntry = result.hooks.PostToolUse.find((m: ClaudeHookMatcher) => m.matcher === 'Skill');
+      expect(skillEntry!.hooks[0].command).toContain('--matcher Skill');
     });
 
     it('does not include --matcher for wildcard entries', async () => {
