+++ src/__tests__/hooks-e2e.test.ts
@@ -39,7 +39,7 @@ async function readResult(filePath: string): Promise<Record<string, unknown>> {
 
 describe('hooks E2E — real file I/O', () => {
   describe('inject — full injection to temp directories', () => {
-    it('creates Claude settings.json with all 4 events and 10 dispatch hooks', async () => {
+    it('creates Claude settings.json with all 4 events and 6 dispatch hooks', async () => {
       const p = claudePath();
       await injectHooks(p, 'claude');
 
@@ -49,11 +49,11 @@ describe('hooks E2E — real file I/O', () => {
       expect(Object.keys(hooks)).toEqual(['SessionStart', 'Stop', 'PostToolUse', 'UserPromptSubmit']);
       expect(hooks.SessionStart).toHaveLength(1);
       expect(hooks.Stop).toHaveLength(1);
-      expect(hooks.PostToolUse).toHaveLength(7);
+      expect(hooks.PostToolUse).toHaveLength(3);
       expect(hooks.UserPromptSubmit).toHaveLength(1);
     });
 
-    it('creates Cursor hooks.json with all 4 events and 10 dispatch hooks', async () => {
+    it('creates Cursor hooks.json with all 4 events and 6 dispatch hooks', async () => {
       const p = cursorPath();
       await injectHooks(p, 'cursor');
 
@@ -64,7 +64,7 @@ describe('hooks E2E — real file I/O', () => {
       expect(Object.keys(hooks)).toEqual(['sessionStart', 'stop', 'postToolUse', 'beforeSubmitPrompt']);
       expect(hooks.sessionStart).toHaveLength(1);
       expect(hooks.stop).toHaveLength(1);
-      expect(hooks.postToolUse).toHaveLength(7);
+      expect(hooks.postToolUse).toHaveLength(3);
       expect(hooks.beforeSubmitPrompt).toHaveLength(1);
     });
