+++ src/builtin-rules.ts
@@ -13,11 +13,12 @@ import fs from 'node:fs/promises';
 //  maintained alongside the CLI code and deployed automatically
 //  on each `teamai pull`.
 //
-//  Currently no built-in rules are deployed — auto-recall hooks
-//  replaced the old teamai-recall.md rule. The infrastructure
-//  is kept for future built-in rules.
-//
-//  Legacy cleanup: removes old teamai-recall.md files on pull.
+//  teamai-recall.md instructs the AI to proactively search the team
+//  knowledge base (via the `teamai-recall` subagent or `teamai recall`)
+//  before starting a task — this replaced the old passive auto-recall
+//  PostToolUse hook, which fired implicitly on every Bash/Grep/WebSearch/
+//  WebFetch call but added noise without the benefit of the subagent's
+//  codebase-graph drill-down and compact structured output.
 //
 
 /** Names of CLI built-in rules. Used by push to exclude them from team repo push. */
@@ -123,10 +124,5 @@ teamai recall "<关键词1> <关键词2> ..."
 \`\`\`
 
 如无相关命中则声明空列表：\`<!-- teamai:referenced-doc-ids: [] -->\`
-
-## 自动错误检索
-
-当 Bash 命令执行报错时，teamai 的 auto-recall hook 会自动搜索团队知识库
-并将相关经验注入上下文。无需手动操作。
 `;
