+++ src/index.ts
@@ -553,18 +553,6 @@ program
   });
 
 program
-  .command('auto-recall', { hidden: true })
-  
-  .description('Auto-recall team knowledge on tool errors (called by PostToolUse hook)')
-  .option('--stdin', 'Read hook data from STDIN')
-  .action(async (cmdOpts) => {
-    if (cmdOpts.stdin) {
-      const { autoRecall } = await import('./auto-recall.js');
-      await autoRecall();
-    }
-  });
-
-program
   .command('todowrite-hint', { hidden: true })
   
   .description('Remind the agent to invoke teamai-recall when TodoWrite is used (PostToolUse hook)')
